requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorlog==6.6.0
cowsay==4.0
emoji==1.7.0
fortune==1.1.0
frozenlist==1.3.0
grizzled-python==2.2.0
hikari==2.0.0.dev108
hikari-lightbulb==2.2.1
idna==3.3
langcodes==3.3.0
language-data==1.1
marisa-trie==0.7.7
multidict==6.0.2
python-dotenv==0.20.0
regex==2022.3.15
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
uvloop==0.16.0; sys_platform != "win32"
yarl==1.7.2
